worked on issue 19 and refactored
