+++ components/style/index.js
@@ -1,0 +1,105 @@
+import {
+  Grid,
+  Card,
+  CardHeader,
+  CardContent,
+  TextField,
+  Button,
+  Typography,
+  Tooltip,
+  MenuItem,
+  Select,
+  InputLabel,
+  FormControl,
+} from "@mui/material";
+import { styled } from "@mui/material/styles";
+
+const StyledCard = styled(Card)(({ theme }) => ({
+  background: "linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%)",
+  boxShadow: "0 3px 5px 2px rgba(255, 105, 135, .3)",
+  borderRadius: "12px",
+  width: "100%",
+  height: "100%",
+  margin: "auto",
+  display: "flex",
+  flexDirection: "column",
+}));
+
+const StyledCardHeader = styled(CardHeader)(({ theme }) => ({
+  textAlign: "center",
+  color: "#fff",
+}));
+
+const StyledCardContent = styled(CardContent)(({ theme }) => ({
+  flexGrow: 1,
+  display: "flex",
+  flexDirection: "column",
+  justifyContent: "space-between",
+}));
+
+const StyledGrid = styled(Grid)(({ theme }) => ({
+  height: "100vh",
+}));
+
+const StyledTypography = styled(Typography)(({ theme }) => ({
+  color: "#fff",
+  fontWeight: "bold",
+}));
+
+const StyledTextField = styled(TextField)(({ theme }) => ({
+  "& label": {
+    color: "#fff",
+    fontWeight: "bold",
+  },
+  "& input": {
+    color: "#fff",
+    fontWeight: "bold",
+  },
+  "& .MuiInput-underline:before": {
+    borderBottomColor: "#fff",
+  },
+}));
+
+const StyledButton = styled(Button)(({ theme }) => ({
+  background: "#fff",
+  color: "#FF8E53",
+  fontWeight: "bold",
+  "&:hover": {
+    background: "#FF8E53",
+    color: "#fff",
+  },
+}));
+
+const StyledFormControl = styled(FormControl)(({ theme }) => ({
+  minWidth: "300px",
+  "& .MuiSelect-select": {
+    paddingRight: theme.spacing(4),
+  },
+  "& .MuiSelect-icon": {
+    right: 0,
+  },
+}));
+
+const StyledSelect = styled(Select)(({ theme }) => ({
+  "&.MuiSelect-select": {
+    paddingRight: theme.spacing(2),
+    "&:focus": {
+      backgroundColor: "transparent",
+    },
+  },
+  "& .MuiSelect-icon": {
+    color: theme.palette.secondary.main,
+  },
+}));
+
+export const StyledComponents = {
+  StyledCard,
+  StyledCardHeader,
+  StyledCardContent,
+  StyledGrid,
+  StyledTypography,
+  StyledTextField,
+  StyledButton,
+  StyledFormControl,
+  StyledSelect,
+};
