+++ pages/index.js
@@ -4,7 +4,6 @@ import { Inter } from "next/font/google";
 import styles from "@/styles/Home.module.css";
 import Link from "next/link";
 
-
 import CheckCircleIcon from "@mui/icons-material/CheckCircle";
 import HourglassEmptyIcon from "@mui/icons-material/HourglassEmpty";
 import ErrorIcon from "@mui/icons-material/Error";
@@ -71,17 +70,15 @@ export default function Home() {
           {/* GitHub */}
           <div>
             <Link
-              href='https://github.com/ank1traj/testcase-generator'
-              rel='noopener noreferrer'
+              href="https://github.com/ank1traj/testcase-generator"
+              rel="noopener noreferrer"
             >
               <code className={styles.code}>Contribute &hearts;</code>
             </Link>
           </div>
         </nav>
         <div className={styles.description}>
-          <p>
-            An Open Source Comprehensive Test Case Generator
-          </p>
+          <p>An Open Source Comprehensive Test Case Generator</p>
           <div className={styles.author}>
             <a
               href="https://www.linkedin.com/in/ank1traj/"
